# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "L5 OpenApi",
        "description": "L5 Swagger OpenApi description",
        "contact": {
            "email": "alexander.gallardo@itpeoplesas.net"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "https://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0.0",
        "x-logo": {
            "url": "https://via.placeholder.com/190x90.png?text=L5-Swagger"
        }
    },
    "servers": [
        {
            "url": "http://localhost:8000",
            "description": "Local Dev Server"
        }
    ],
    "paths": {
        "/api/lead": {
            "post": {
                "tags": [
                    "Lead"
                ],
                "summary": "Add Lead",
                "description": "Add Lead",
                "operationId": "createLead",
                "requestBody": {
                    "description": "Lead object that needs to be created",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "name": {
                                        "description": "Name",
                                        "type": "string"
                                    },
                                    "source": {
                                        "description": "Source",
                                        "type": "string"
                                    },
                                    "owner": {
                                        "description": "Owner",
                                        "type": "integer"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Ok"
                    },
                    "401": {
                        "description": "Unauthorized"
                    }
                }
            }
        },
        "/api/leads": {
            "get": {
                "tags": [
                    "Lead"
                ],
                "summary": "Show all leads",
                "description": "Return all leads",
                "operationId": "getLeadsList",
                "responses": {
                    "200": {
                        "description": "Ok"
                    },
                    "401": {
                        "description": "Unauthorized"
                    }
                }
            }
        },
        "/api/lead/{id}": {
            "get": {
                "tags": [
                    "Lead"
                ],
                "summary": "Show by ID",
                "description": "Retorn lead",
                "operationId": "searchLeadById",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID Lead",
                        "required": true,
                        "schema": {
                            "type": "string"
                        },
                        "examples": {
                            "int": {
                                "summary": "ID Lead",
                                "value": "6372930455a84eb00e0e5852"
                            }
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Ok"
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "404": {
                        "description": "Not found"
                    }
                }
            }
        },
        "/api/auth/auth": {
            "post": {
                "tags": [
                    "Auth"
                ],
                "summary": "Get Token",
                "description": "Get Token with username and password",
                "operationId": "login",
                "requestBody": {
                    "description": "Get Token with username and password",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "username": {
                                        "description": "Username",
                                        "type": "string"
                                    },
                                    "password": {
                                        "description": "Password",
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Ok"
                    },
                    "401": {
                        "description": "Unauthorized"
                    }
                }
            }
        }
    },
    "components": {
        "securitySchemes": {
            "bearer": {
                "type": "http",
                "description": "Login with username and password to get the authentication token",
                "in": "header",
                "scheme": "bearer"
            }
        }
    },
    "security": [
        {
            "bearer": []
        }
    ]
}